+++ CHANGELOG.md
@@ -11,9 +11,11 @@
 - Peerless Skill (Modern)
 - Warrior of the Gods (Legacy)
 - Channel Divinity (Modern)
+- Shift: Wildhunt (Legacy)
 ## Update Notes:
 - This update mostly consists of making older automations automatically apply missing scale values as well as adding more self-correcting resource consumption linking.
 - Added a setting to manual rolls to use dice totals instead of overall totals. (Thanks Tyler)
+- It is a known issue that Psi-Bolstered Knack isn't displaying changes to tool checks in the chat card. The next update to Midi-Qol should fix this.
 
 # 1.3.56 Release:
 ## New Modern Content: